requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
bunch==1.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
importlib-resources==5.7.1
ipfshttpclient==0.8.0a2
jsonschema==4.4.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
munch==2.5.0
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.20.1
pycryptodome==3.14.1
pyrsistent==0.18.1
python-dotenv==0.20.0
requests==2.27.1
rlp==2.0.1
six==1.16.0
tinydb==4.7.0
toolz==0.11.2
urllib3==1.26.9
varint==1.0.2
web3==5.29.0
websockets==9.1
wincertstore==0.2
yarl==1.7.2
zipp==3.8.0
